fix(blog): flatten article grid to avoid hydration layout shift
ArticlePreview chunked posts into rows based on useMediaQuery, which
returns false on SSR and on the first client render. On /blog the
server-rendered HTML therefore had 12 single-card rows, and desktop
visitors on slow networks saw 1 card above the fold until matchMedia
resolved and the grid re-chunked to 3 per row.

Render all posts inside a single motion.ul with the responsive CSS
Grid handling columns. SSR output matches the hydrated DOM, so there
is nothing to shift. Also drops the now-unused ArticlePreviewRow and
the hasFixedItemsPerRow prop from LatestFromOurBlog.
