+++ src/components/ArticlePreview/ArticlePreview.tsx
@@ -1,50 +1,64 @@
 import React, { FC } from 'react';
-import { chunk, isEmpty } from 'lodash';
-import { useMediaQuery } from 'react-responsive';
-import { BlogPostDto, MEDIA_DESKTOP_SM, MEDIA_TABLET_SM, PropsWithAnimation } from '@app/utils';
+import { isEmpty } from 'lodash';
+import { motion } from 'framer-motion';
+import {
+  BlogPostDto,
+  createBemBlockBuilder,
+  getEaseInOutTransition,
+  opacityScaleAnimationProps,
+  PropsWithAnimation,
+} from '@app/utils';
+import { useInView } from '@app/hooks/useInView';
+import { useMotionEnterAnimation } from '@app/hooks/useMotionEnterAnimation';
 
-import { ArticlePreviewRow } from './ArticlePreviewRow';
+import { ArticlePreviewItem } from './ArticlePreviewItem';
 
 import './ArticlePreview.scss';
 
 interface ArticlePreviewProps {
   posts: BlogPostDto[];
-  hasFixedItemsPerRow?: boolean;
   searchQuery?: string;
 }
 
-const getItemsPerRow = (isDesktop: boolean, isTablet: boolean) => {
-  if (isDesktop) {
-    return 3;
-  }
-
-  if (isTablet) {
-    return 2;
-  }
-
-  return 1;
-};
+const getBlocksWith = createBemBlockBuilder(['article-preview-list']);
 
 export const ArticlePreview: FC<PropsWithAnimation<ArticlePreviewProps>> = ({
   posts,
-  hasFixedItemsPerRow = false,
   isAnimationEnabled = false,
   searchQuery,
 }) => {
-  const isDesktop = useMediaQuery({ query: MEDIA_DESKTOP_SM });
-  const isTablet = useMediaQuery({ query: MEDIA_TABLET_SM });
-  const rows = chunk(posts, hasFixedItemsPerRow ? 3 : getItemsPerRow(isDesktop, isTablet));
+  const [listRef, isInView] = useInView();
+  const getAnimation = useMotionEnterAnimation(
+    {
+      ...opacityScaleAnimationProps,
+      ...getEaseInOutTransition(0.7),
+    },
+    isAnimationEnabled,
+  );
 
-  return !isEmpty(posts) ? (
-    <ul>
-      {rows.map((row, rowIndex) => (
-        <ArticlePreviewRow
-          key={rowIndex}
-          row={row}
-          isAnimationEnabled={isAnimationEnabled}
-          searchQuery={searchQuery}
-        />
+  if (isEmpty(posts)) {
+    return null;
+  }
+
+  return (
+    <motion.ul
+      ref={listRef}
+      className={getBlocksWith()}
+      {...getAnimation({
+        isInView,
+        additionalEffects: {
+          hiddenAdditional: {
+            y: 150,
+          },
+          enterAdditional: {
+            y: 0,
+          },
+        },
+      })}
+    >
+      {posts.map(post => (
+        <ArticlePreviewItem key={post.id} post={post} searchQuery={searchQuery} />
       ))}
-    </ul>
-  ) : null;
+    </motion.ul>
+  );
 };
