--- src/components/ArticlePreview/ArticlePreviewRow/index.ts
@@ -1,1 +1,0 @@
-export * from './ArticlePreviewRow';
